feat: birdhouseAI page mavbar
