+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "cSpell.words": ["Agencidev", "agencydev", "personalised"]
+  "cSpell.words": ["Agencidev", "agencydev", "favourite", "personalised"]
 }
