+++ app/_components/birdhouseNav/birdhouseNav.module.css
@@ -1,8 +1,39 @@
+.navbarContainer {
+  position: fixed;
+  top: 0;
+  width: 100%;
+  z-index: 1000;
+  display: flex;
+  justify-content: space-between;
+  transition: all 0.3s ease;
+  background-color: rgba(255, 255, 255, 1);
+  backdrop-filter: blur(5px);
+  z-index: 1000;
+  padding: 20px 40px;
+}
+
 .logoWrapper {
   display: flex;
-  justify-content: center;
   position: relative;
   width: 100%;
   height: 200px;
   transition: all 0.3s ease;
+}
+
+.navLinks {
+  display: flex;
+  transition: all 0.3s ease;
+  color: var(--text);
+  cursor: pointer;
+}
+
+.navLinks li {
+  transition: all 0.2s ease;
+  position: relative;
+  color: var(--text);
+}
+
+.navLinks li:hover {
+  transform: scale(1.1);
+  opacity: 0.8;
 }
