+++ app/birdhouseAi/page.tsx
@@ -1,9 +1,12 @@
-import React from 'react'
+import React from "react";
+import BirdhouseNav from "../_components/birdhouseNav/BirdhouseNav";
 
 const page = () => {
   return (
-    <div>page</div>
-  )
-}
+    <div>
+      <BirdhouseNav />
+    </div>
+  );
+};
 
-export default page
+export default page;
